residual_fear/requirements.txt
openai
requests
python-dotenv
numpy
hbdscan
opencv-python
insightface
openai-whisper
google-api-python-client
google-auth
google-auth-oauthlib
